+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Forecast datasource API
-- GFS Forecast datasource
+- GFS forecast datasource
+- GEFS forecast datasource
 
 ### Changed